src/main/java/com/xschen/works/gp/lgp/program/operators/Sqrt.java
package com.xschen.works.gp.lgp.program.operators;


import com.xschen.works.gp.lgp.enums.OperatorExecutionStatus;
import com.xschen.works.gp.lgp.program.Operator;
import com.xschen.works.gp.lgp.program.ProgramManager;
import com.xschen.works.gp.lgp.program.Register;


/**
 * Created by xschen on 30/4/2017.
 */
public class Sqrt extends Operator {

   private static final long serialVersionUID = -3206913752183532206L;


   public Sqrt(){
      super("^");
   }

   @Override public Operator makeCopy() {
      Sqrt clone = new Sqrt();
      clone.copy(this);
      return clone;
   }


   @Override public OperatorExecutionStatus execute(Register operand1, Register operand2, Register destination_register) {
      double x = operand1.getValue();
      destination_register.setValue(Math.sqrt(Math.abs(x)));
      return OperatorExecutionStatus.LGP_EXECUTE_NEXT_INSTRUCTION;
   }
}
